Create recv_exactly function Instead of bufferedIO
Works better with py2.7 and timeouts of underlying sockets.

+++ umodbus/functions.py
@@ -55,6 +55,7 @@ A response PDU could look like this::
     'b\\x06'
 
 """
+from __future__ import division
 import struct
 import inspect
 import math
@@ -280,7 +281,7 @@ class ReadCoils(ModbusFunction):
 
         :return: number of bytes.
         """
-        return 2 + math.ceil(self.quantity / 8)
+        return 2 + int(math.ceil(self.quantity / 8))
 
     def create_response_pdu(self, data):
         """ Create response pdu.
@@ -493,7 +494,7 @@ class ReadDiscreteInputs(ModbusFunction):
 
         :return: number of bytes.
         """
-        return 2 + math.ceil(self.quantity / 8)
+        return 2 + int(math.ceil(self.quantity / 8))
 
     def create_response_pdu(self, data):
         """ Create response pdu.
